+++ src/components/Modal/Modal.js
@@ -1,0 +1,31 @@
+import React, { Component } from 'react';
+
+class Modal extends Component {
+  render() {
+    <div
+      className="modal-content"
+    >
+      <div
+        className="modal-header"
+      >
+        Test THIS
+      </div>
+      <div
+        className="modal-body"
+      >
+        <p>OMG ITS A MODAL</p>
+      </div>
+      <div
+        className="modal-footer"
+      >
+        <button
+          onClick={click}
+        >
+          x
+        </button>
+      </div>
+    </div>
+  }
+}
+
+export default Modal;
